Refactor code for getting asana data from using partial application to using event listeners.
The code to get data from asana initially used nested callbacks, made easier with partial application.
This was a nightmare to understand, so I refactored using event emitters and listerners as
recommended in http://wekeroad.com/2012/04/05/cleaning-up-deep-callback-nesting-with-nodes-eventemitter
I also added saving asana data to the database in a collection called 'projects'.