Refactor DataGetter so that it doesn't save the data and so it takes an auth parameter.
Removed saving functions from DataGetter because a) separation of concerns and b) we
don't need to save anything to the database anyway - we can just push straight to
backbone.

Made DataGetter take an auth parameter so that every dataGetter object constructed from
DataGetter will have an auth paramter in it. This is because the user won't change, so all
we have to pass in is the path required.