DRY down the long list of app.gets to just one app.get('/data/*'
Change the long list of get listeners to just one listening for '/data/*'
This is so that there is less code.
Used '/data/*' rather than just '/*' because we want routes like 'js/main.js'
to be handled differently.

Also changed the backbone collection routes to reflect this change to /data/route