Add getDeepData public method DataGetter constructor.
Added getDeepData method to DataGetter constructor. Previously,
it just had the getData method which would get JSON data from, for
example, /users. But more detailed information is availiable for each
user by going to /users/id for each user. So getDeepData does this
and emits a "gotDeepData" even along with a list that contains
the deep data.

Also altered the app.get('/' ... listener so that deep data is sent
to the jade template, rather than shallow data.